Update to Xcode installation instructions
Thanks for putting this page together! 🙂  On the Xcode setup section I wondered if it would be helpful to link to the iOS app repository with our setup guide. We keep this up to date with the installation process and other steps required to get the app up and running.